+++ src/main/java/Hachiware/AddCommand.java
@@ -1,16 +1,14 @@
-public class TodoCommand extends Command {
-    private final String description;
+package Hachiware;
 
-    public TodoCommand(String description) throws HachiwareException {
-        if (description.isEmpty()) {
-            throw new HachiwareException("MEOW!!! OI The description of a todo cannot be empty.");
-        }
-        this.description = description;
+public class AddCommand extends Command {
+    private final Task task;
+
+    public AddCommand(Task task) {
+        this.task = task;
     }
 
     @Override
     public void execute(TaskList tasks, Ui ui, StoreFile storage) throws HachiwareException {
-        Task task = new ToDo(description);
         tasks.add(task);
         storage.save(tasks.getAll());
         ui.printLine();
@@ -24,4 +22,5 @@ public class TodoCommand extends Command {
     public boolean isExit() {
         return false;
     }
+
 }
