+++ src/main/java/Hachiware/ByeCommand.java
@@ -1,3 +1,5 @@
+package Hachiware;
+
 public class ByeCommand extends Command {
     @Override
     public void execute(TaskList tasks, Ui ui, StoreFile storage) {
