+++ src/main/java/Hachiware/Event.java
@@ -1,4 +1,5 @@
-import java.time.LocalDate;
+package Hachiware;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
