+++ src/main/java/Hachiware/ExitCommand.java
@@ -1,3 +1,5 @@
+package Hachiware;
+
 public class ExitCommand extends Command {
     @Override
     public void execute(TaskList tasks, Ui ui, StoreFile storeFile) {
