+++ src/main/java/Hachiware/HachiwareException.java
@@ -1,3 +1,5 @@
+package Hachiware;
+
 public class HachiwareException extends Exception {
 
     public HachiwareException(String message) {
