+++ src/main/java/Hachiware/ToDo.java
@@ -1,0 +1,14 @@
+package Hachiware;
+
+public class ToDo extends Task {
+
+    public ToDo(String description) {
+        super(description);
+    }
+
+    @Override
+    public String toString() {
+        return "[T]" + super.toString();
+    }
+
+}
