+++ src/main/java/Hachiware/TodoCommand.java
@@ -1,12 +1,18 @@
-public class AddCommand extends Command {
-    private final Task task;
+package Hachiware;
 
-    public AddCommand(Task task) {
-        this.task = task;
+public class TodoCommand extends Command {
+    private final String description;
+
+    public TodoCommand(String description) throws HachiwareException {
+        if (description.isEmpty()) {
+            throw new HachiwareException("MEOW!!! OI The description of a todo cannot be empty.");
+        }
+        this.description = description;
     }
 
     @Override
     public void execute(TaskList tasks, Ui ui, StoreFile storage) throws HachiwareException {
+        Task task = new ToDo(description);
         tasks.add(task);
         storage.save(tasks.getAll());
         ui.printLine();
@@ -20,5 +26,4 @@ public class AddCommand extends Command {
     public boolean isExit() {
         return false;
     }
-
 }
