+++ src/main/java/Hachiware/UnmarkCommand.java
@@ -1,7 +1,9 @@
-public class MarkCommand extends Command {
+package Hachiware;
+
+public class UnmarkCommand extends Command {
     private final int index;
 
-    public MarkCommand(String arg) throws HachiwareException {
+    public UnmarkCommand(String arg) throws HachiwareException {
         try {
             this.index = Integer.parseInt(arg.trim()) - 1;
         } catch (NumberFormatException e) {
@@ -12,13 +14,13 @@ public class MarkCommand extends Command {
     @Override
     public void execute(TaskList tasks, Ui ui, StoreFile storage) throws HachiwareException {
         if (index < 0 || index >= tasks.size()) {
-            throw new HachiwareException("Task number out of bounds");
+            throw new HachiwareException("Hachiware.Hachiware.Task number out of bounds");
         }
         Task task = tasks.get(index);
-        task.markAsDone();
+        task.markAsNotDone();
         storage.save(tasks.getAll());
         ui.printLine();
-        System.out.println("Nice! I've marked this task as done: ");
+        System.out.println("OK, I've marked this task as not done yet: ");
         System.out.println(task);
         ui.printLine();
     }
